+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
 
+### Added
+
+- react hook form storybook
+
 ### Fixed
 
 - DatePickerInput null value not reflected to display